Hide popup extension widget before unregister (#1)
* Mavenize - create from archetype

* Mavenize - reorganize files

* Hide popup before removing connector (#16)

* Clean readme file
