+++ README.md
@@ -1,0 +1,1 @@
+# PopupExtension Add-on for Vaadin 7
