+++ popupextension-addon/src/main/java/org/vaadin/addons/client/PopupExtensionDataTransferComponentWidget.java
@@ -1,0 +1,10 @@
+package org.vaadin.addons.client;
+
+import com.google.gwt.user.client.DOM;
+import com.google.gwt.user.client.ui.Widget;
+
+public class PopupExtensionDataTransferComponentWidget extends Widget {
+    public PopupExtensionDataTransferComponentWidget() {
+        setElement(DOM.createDiv());
+    }
+}
