+++ popupextension-addon/src/main/java/org/vaadin/addons/client/PopupExtensionServerRpc.java
@@ -1,0 +1,7 @@
+package org.vaadin.addons.client;
+
+import com.vaadin.shared.communication.ServerRpc;
+
+public interface PopupExtensionServerRpc extends ServerRpc {
+    void setOpen(boolean open);
+}
