+++ popupextension-addon/src/main/resources/org/vaadin/addons/public/popupextension/styles.css
