+++ popupextension-addon/src/test/java/org/vaadin/addons/MyComponentTest.java
@@ -1,0 +1,13 @@
+package org.vaadin.addons;
+
+import junit.framework.Assert;
+import org.junit.Test;
+
+// JUnit tests here
+public class MyComponentTest {
+
+	@Test
+	public void thisAlwaysPasses() {
+		Assert.assertEquals(true, true);
+	}
+}
