+++ popupextension-demo/src/main/resources/org/vaadin/addons/README
@@ -1,0 +1,1 @@
+Please add your static resources here
