--- src/com/github/wolfie/popupextension/PopupextensionWidgetset.gwt.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE module PUBLIC "-//Google Inc.//DTD Google Web Toolkit 1.7.0//EN" "http://google-web-toolkit.googlecode.com/svn/tags/1.7.0/distro-source/core/src/gwt-module.dtd">
-<module>
-	<inherits name="com.vaadin.DefaultWidgetSet" />
-	<stylesheet src="popupextension.css"/>
-</module>
