--- src/com/github/wolfie/popupextension/client/PopupExtensionDataTransferComponentState.java
@@ -1,9 +1,0 @@
-package com.github.wolfie.popupextension.client;
-
-import com.vaadin.shared.AbstractComponentState;
-
-public class PopupExtensionDataTransferComponentState extends
-		AbstractComponentState {
-	private static final long serialVersionUID = 6235786382671700479L;
-	public String popupId;
-}
