--- src/com/github/wolfie/popupextension/client/PopupExtensionDataTransferComponentWidget.java
@@ -1,10 +1,0 @@
-package com.github.wolfie.popupextension.client;
-
-import com.google.gwt.user.client.DOM;
-import com.google.gwt.user.client.ui.Widget;
-
-public class PopupExtensionDataTransferComponentWidget extends Widget {
-	public PopupExtensionDataTransferComponentWidget() {
-		setElement(DOM.createDiv());
-	}
-}
