--- src/com/github/wolfie/popupextension/client/PopupExtensionServerRpc.java
@@ -1,7 +1,0 @@
-package com.github.wolfie.popupextension.client;
-
-import com.vaadin.shared.communication.ServerRpc;
-
-public interface PopupExtensionServerRpc extends ServerRpc {
-	void setOpen(boolean open);
-}
